Fixed sign on negative float numbers greater than -1
As reported by terasaka2k:

"%.2f".printf(-0.33) generates "0.33".

+++ src/sprintf.js
@@ -78,7 +78,7 @@
                 /* force sign */
                 sign: ( part[1] == '+' ),
                 /* is the given data negative */
-                negative: ( parseInt( arguments[paramIndex] ) < 0 ) ? true : false,
+                negative: ( parseFloat( arguments[paramIndex] ) < 0 ) ? true : false,
                 /* padding character (default: <space>) */
                 padding: ( part[2] == undefined )
                          ? ( ' ' ) /* default */
